Performance-optimized version of the stopwatch ticker UI
# Conflicts:
#	projects/stopwatch_flutter/lib/ui/stopwatch.dart
#	projects/stopwatch_flutter/lib/ui/stopwatch_renderer.dart
