+++ projects/stopwatch_flutter/lib/ui/stopwatch.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/scheduler.dart';
 import 'package:stopwatch_flutter/ui/reset_button.dart';
 import 'package:stopwatch_flutter/ui/start_stop_button.dart';
 import 'package:stopwatch_flutter/ui/stopwatch_renderer.dart';
+import 'package:stopwatch_flutter/ui/stopwatch_ticker_ui.dart';
 
 class Stopwatch extends StatefulWidget {
   const Stopwatch({super.key});
@@ -11,49 +11,23 @@ class Stopwatch extends StatefulWidget {
   State<Stopwatch> createState() => _StopwatchState();
 }
 
-class _StopwatchState extends State<Stopwatch>
-    with SingleTickerProviderStateMixin {
-  Duration _previouslyElapsed = Duration.zero;
-  Duration _currentlyElapsed = Duration.zero;
-  Duration get _elapsed => _previouslyElapsed + _currentlyElapsed;
+class _StopwatchState extends State<Stopwatch> {
+  /// Global key used to manipulate the state of the StopwatchTickerUI
+  final _tickerUIKey = GlobalKey<StopwatchTickerUIState>();
   bool _isRunning = false;
-  late final Ticker _ticker;
-  @override
-  void initState() {
-    super.initState();
-    _ticker = createTicker((elapsed) {
-      setState(() {
-        _currentlyElapsed = elapsed;
-      });
-    });
-  }
-
-  @override
-  void dispose() {
-    _ticker.dispose();
-    super.dispose();
-  }
 
   void _toggleRunning() {
     setState(() {
       _isRunning = !_isRunning;
-      if (_isRunning) {
-        _ticker.start();
-      } else {
-        _ticker.stop();
-        _previouslyElapsed += _currentlyElapsed;
-        _currentlyElapsed = Duration.zero;
-      }
     });
+    _tickerUIKey.currentState?.toggleRunning(_isRunning);
   }
 
   void _reset() {
-    _ticker.stop();
     setState(() {
       _isRunning = false;
-      _previouslyElapsed = Duration.zero;
-      _currentlyElapsed = Duration.zero;
     });
+    _tickerUIKey.currentState?.reset();
   }
 
   @override
@@ -63,10 +37,14 @@ class _StopwatchState extends State<Stopwatch>
         final radius = constraints.maxWidth / 2;
         return Stack(
           children: [
-            StopwatchRenderer(
-              elapsed: _elapsed,
+            // non-ticker dependent UI
+            StopwatchRenderer(radius: radius),
+            // ticker dependent UI
+            StopwatchTickerUI(
+              key: _tickerUIKey,
               radius: radius,
             ),
+            // reset button
             Align(
               alignment: Alignment.bottomLeft,
               child: SizedBox(
@@ -77,6 +55,7 @@ class _StopwatchState extends State<Stopwatch>
                 ),
               ),
             ),
+            // start/stop button
             Align(
               alignment: Alignment.bottomRight,
               child: SizedBox(
