+++ projects/stopwatch_flutter/lib/ui/stopwatch_renderer.dart
@@ -1,17 +1,11 @@
-import 'dart:math';
-
 import 'package:flutter/material.dart';
-import 'package:stopwatch_flutter/ui/clock_hand.dart';
 import 'package:stopwatch_flutter/ui/clock_markers.dart';
-import 'package:stopwatch_flutter/ui/elapsed_time_text.dart';
 
 class StopwatchRenderer extends StatelessWidget {
   const StopwatchRenderer({
     super.key,
-    required this.elapsed,
     required this.radius,
   });
-  final Duration elapsed;
   final double radius;
 
   @override
@@ -30,24 +24,6 @@ class StopwatchRenderer extends StatelessWidget {
             left: radius,
             child: ClockTextMarker(value: i, maxValue: 60, radius: radius),
           ),
-        Positioned(
-          left: radius,
-          top: radius,
-          child: ClockHand(
-            handLength: radius,
-            handThickness: 2,
-            rotationZAngle: pi + (2 * pi / 60000) * elapsed.inMilliseconds,
-            color: Colors.orange,
-          ),
-        ),
-        Positioned(
-          left: 0,
-          right: 0,
-          top: radius * 1.3,
-          child: ElapsedTimeText(
-            elapsed: elapsed,
-          ),
-        ),
       ],
     );
   }
